Module google_servicebroker1::api

Structs

GoogleIamV1__Binding
Associates members with a role.

GoogleIamV1__Policy
Defines an Identity and Access Management (IAM) policy. It is used to specify access control policies for Cloud Platform resources.

GoogleIamV1__SetIamPolicyRequest
Request message for SetIamPolicy method.

GoogleIamV1__TestIamPermissionsRequest
Request message for TestIamPermissions method.

GoogleIamV1__TestIamPermissionsResponse
Response message for TestIamPermissions method.

GoogleType__Expr
Represents an expression text. Example:

MethodGetIamPolicyCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

MethodMethods
A builder providing access to all free methods, which are not associated with a particular resource. It is not used directly, but through the ServiceBroker hub.

MethodSetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy.

MethodTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.

ServiceBroker
Central instance to access all ServiceBroker related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.